// frontend/src/api/generated/models/tradeSortKey.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fusion Trade Blotter API
 * OpenAPI spec version: 1.0.0
 */

export type TradeSortKey = typeof TradeSortKey[keyof typeof TradeSortKey];


export const TradeSortKey = {
  tradeId: 'tradeId',
  symbol: 'symbol',
  side: 'side',
  quantity: 'quantity',
  price: 'price',
  trader: 'trader',
  book: 'book',
  counterparty: 'counterparty',
  tradeTimestamp: 'tradeTimestamp',
  status: 'status',
  version: 'version',
  notional: 'notional',
  signedQuantity: 'signedQuantity',
  signedNotional: 'signedNotional',
} as const;